Added task2 of cg-course

+++ src/Render2D/src/Renderer2D.cpp
@@ -25,8 +25,8 @@ void Renderer2D::SetViewport(Vector2u const& newSize)
 void Renderer2D::BeginScene(Vector2f const& cameraPos, const float cameraZoom)
 {
 	const Vector2f worldSize = {
-		m_viewportSize.x / cameraZoom,
-		m_viewportSize.y / cameraZoom
+		static_cast<float>(m_viewportSize.x) / cameraZoom,
+		static_cast<float>(m_viewportSize.y) / cameraZoom
 	};
 
 	m_api->SetCamera(cameraPos, worldSize);
